V2.0.1
Reformat FiberPhix into a singular package containing all functions.
1) FiberPhixDAT: replaced with FiberPhix.streams() call.
2) FiberPhixDAT_Epocs: replaced with FiberPhix.epocs().
3)FiberPhixDAT_MainInfo: replaced with FiberPhix.getinfo()
4)FiberPhixDAT_BatchMainInfo: Replaced with FiberPhix.batchinfo()

Other notable improvements:
1) Methods are easily shared and imported for future use
2) Significantly improved endpoint user workflow (single call vs. multiple calls)
3) User values spreasheet has been updated and now requests additional information (figure saving path, information saving path, as well as batch save location)
4) Uniform syntax across scripts
